+++ test/correctness/PySys_internal_076/run.py
@@ -13,6 +13,7 @@ class PySysTest(BaseTest):
 	def execute(self):
 		self.copy(self.input, self.output+'/test')
 		runPySys(self, 'pysys', ['run', '-o', self.output+'/output', '--purge', '-v', 'debug'], workingDir='test')
+		runPySys(self, 'pysys-preserveEmptyOutputs', ['run', '-o', self.output+'/output-preserveEmptyOutputs', '--preserveEmptyOutputs', '-v', 'debug'], workingDir='test')
 	
 			
 	def validate(self):
@@ -25,12 +26,18 @@ class PySysTest(BaseTest):
 		self.assertPathExists('output/PySys_NestedTestcase/dir1/dir1a/', exists=False) 
 		self.assertPathExists('output/PySys_NestedTestcase/dir2/dir2a', exists=False) 
 		self.assertPathExists('output/PySys_NestedTestcase/dir2/', exists=False)
+		self.assertGrep('pysys.out', ' DEBUG .*Purged empty output directory now that test is complete: .*/PySys_NestedTestcase/dir2'.replace('/', r'[/\\]'))
 
 
 		self.assertPathExists('output/PySys_NestedTestcaseInspect/run.log')
 		self.assertPathExists('output/PySys_NestedTestcaseInspect/nonempty.txt', exists=True)
 		self.assertPathExists('output/PySys_NestedTestcaseInspect/empty.txt', exists=False)
 		
-		self.assertPathExists('output/PySys_NestedTestcaseInspect/dir1/dir1a/', exists=True) # empty but not emptied by PySys
-		self.assertPathExists('output/PySys_NestedTesPySys_NestedTestcaseInspecttcase/dir2/dir2a', exists=False) # emptied by PySys deleting empty file
-		self.assertPathExists('output/PySys_NestedTestcaseInspect/dir2/', exists=True)
+		self.assertPathExists('output/PySys_NestedTestcaseInspect/dir1/dir1a/', exists=False) # empty so deleted by PySys, but with info logging
+
+		self.assertPathExists('output/PySys_NestedTestcaseInspect/dir2/dir2a', exists=False) # emptied by PySys deleting empty file
+		self.assertGrep('pysys.out', ' INFO .*Purged empty output directory now that test is complete: .*/PySys_NestedTestcaseInspect/dir2'.replace('/', r'[/\\]'))
+		self.assertPathExists('output/PySys_NestedTestcaseInspect/dir2/', exists=False)
+
+		self.assertPathExists('output-preserveEmptyOutputs/PySys_NestedTestcase/dir2/dir2a/empty.txt')
+		self.assertPathExists('output-preserveEmptyOutputs/PySys_NestedTestcase/dir1')